+++ src/handlers/html/index.rs
@@ -1,10 +1,18 @@
-use crate::{Highlighter, Page};
+use crate::{handlers::extract::Theme, Highlighter, Page};
 use askama::Template;
 use axum::extract::State;
 
 /// GET handler for the index page.
-pub async fn get(State(page): State<Page>, State(highlighter): State<Highlighter>) -> Index {
-    Index { page, highlighter }
+pub async fn get(
+    State(page): State<Page>,
+    State(highlighter): State<Highlighter>,
+    theme: Option<Theme>,
+) -> Index {
+    Index {
+        page,
+        theme,
+        highlighter,
+    }
 }
 
 /// Index page displaying a form for paste insertion and a selection box for languages.
@@ -12,5 +20,6 @@ pub async fn get(State(page): State<Page>, State(highlighter): State<Highlighter
 #[template(path = "index.html")]
 pub struct Index {
     page: Page,
+    theme: Option<Theme>,
     highlighter: Highlighter,
 }