+++ src/handlers/mod.rs
@@ -1,6 +1,7 @@
 pub mod delete;
 pub mod download;
-mod extract;
+pub mod extract;
 pub mod html;
 pub mod insert;
 pub mod raw;
+pub mod theme;